# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Secure Patrol API
  version: 1.0.0
  description: |
    Dokumentasi API Secure Patrol. **Hanya tersedia di environment development** (`APP_ENV=development`).

    ## Cara mencoba API di halaman ini
    1. Klik **Authorize**, isi **AppId** (`X-App-Id`) dan **AppKey**. App Key hanya dipakai browser untuk
       menandatangani request dan **tidak pernah dikirim** ke server.
    2. Panggil `POST /auth/login`. Access token otomatis dipasang ke **BearerAuth**.
    3. Endpoint lain siap dicoba.

    Belum punya App ID/Key? Jalankan `go run . create-app-client -name "Dev Docs" -platform web`.

    ## Keamanan request
    Setiap request ke `/api/v1` wajib ditandatangani dengan header `X-App-Id`, `X-Timestamp`, `X-Nonce`, dan
    `X-Signature`. Halaman ini membuat ketiga header terakhir secara otomatis.

    ```
    payload   = METHOD \n REQUEST_URI \n TIMESTAMP \n NONCE \n SHA256_HEX(BODY)
    signature = HEX(HMAC_SHA256(APP_KEY, payload))
    ```

    Detail lengkap dan contoh kode mobile ada di `docs/AUTH.md`.

    ## Pembatasan platform
    **Login per platform:** App Client `android` dan `ios` bisa dipakai semua role. App Client `web` (dan `server`)
    hanya untuk Super-Admin, Manager Keamanan, Kepala Keamanan, dan Admin Keamanan; role lain mendapat `403`
    `your role is not allowed to sign in on this platform` saat login, refresh, maupun request berikutnya.

...
    membuat, mengubah, dan menghapus patrol point dan patrol shift.

...

...

...
servers:
...
security:
  - AppId: []
    AppKey: []
    BearerAuth: []
paths:
...
  /roles:
    get:
      operationId: listRoles
      tags: [Roles]
      summary: Daftar role
      description: '**Hanya platform web.** Akses: Super-Admin, Manager Keamanan, Admin Keamanan.'
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Daftar role
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Role' }
                          pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createRole
      tags: [Roles]
      summary: Tambah role
      description: '**Hanya platform web.** Akses: Super-Admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /roles/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getRole
      tags: [Roles]
      summary: Detail role
      description: '**Hanya platform web.** Akses: Super-Admin, Manager Keamanan, Admin Keamanan.'
      responses:
        '200':
          description: Detail role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateRole
      tags: [Roles]
      summary: Ubah role
      description: '**Hanya platform web.** Akses: Super-Admin. Kode role tidak bisa diubah; role sistem tidak bisa dinonaktifkan.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role diubah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      operationId: deleteRole
      tags: [Roles]
      summary: Hapus role (soft delete)
      description: '**Hanya platform web.** Akses: Super-Admin. Role sistem dan role yang masih dipakai user tidak bisa dihapus. Kode role yang dihapus bisa dipakai lagi.'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /patrol-points:
    get:
      operationId: listPatrolPoints
      tags: [Patrol Points]
      summary: Daftar patrol point
      description: 'Semua user yang login, semua platform. Search mencari di nama, lokasi, dan kode NFC.'
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Daftar patrol point, diurutkan berdasarkan nama
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/PatrolPoint' }
                          pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createPatrolPoint
      tags: [Patrol Points]
      summary: Tambah patrol point
      description: |
        **Hanya platform web.** Akses: Super-Admin, Manager Keamanan, Kepala Keamanan, Admin Keamanan.
        Kode NFC disimpan dalam huruf besar dan harus unik.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatrolPointRequest'
      responses:
        '201':
          description: Patrol point dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatrolPointResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /patrol-points/by-nfc:
    get:
      operationId: getPatrolPointByNfc
      tags: [Patrol Points]
      summary: Cari patrol point dari tag NFC yang dipindai
      description: 'Semua user yang login, semua platform. Pencarian tidak membedakan huruf besar/kecil.'
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
          example: '04:A2:1F:9C'
      responses:
        '200':
          description: Patrol point dari tag NFC tersebut
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatrolPointResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /patrol-points/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getPatrolPoint
      tags: [Patrol Points]
      summary: Detail patrol point
      description: 'Semua user yang login, semua platform.'
      responses:
        '200':
          description: Detail patrol point
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatrolPointResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updatePatrolPoint
      tags: [Patrol Points]
      summary: Ubah patrol point
      description: '**Hanya platform web.** Akses: Super-Admin, Manager Keamanan, Kepala Keamanan, Admin Keamanan.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatrolPointRequest'
      responses:
        '200':
          description: Patrol point diubah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatrolPointResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      operationId: deletePatrolPoint
      tags: [Patrol Points]
      summary: Hapus patrol point (soft delete)
      description: '**Hanya platform web.** Kode NFC patrol point yang dihapus bisa dipakai lagi.'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /settings:
    get:
      operationId: listSettings
      tags: [Settings]
      summary: Daftar pengaturan sistem
      description: |
        **Hanya platform web.** Akses baca: Super-Admin, Manager Keamanan, Kepala Keamanan, Admin Keamanan.
        Berisi nilai saat ini, nilai default, tipe, dan batas setiap setting.
      responses:
        '200':
          description: Semua setting
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettingsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      operationId: updateSettings
      tags: [Settings]
      summary: Ubah pengaturan sistem
      description: |
        **Hanya platform web.** Akses: Super-Admin. Ubah satu atau beberapa setting sekaligus; perubahan langsung
        berlaku tanpa restart dan tercatat di audit log. Nilai `null` mengembalikan setting ke default.
        Jika satu nilai tidak valid, tidak ada yang disimpan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [values]
              properties:
                values:
                  type: object
                  additionalProperties: true
                  example: { patrol_location_radius_meters: 150, face_mobile_accuracy: 0.8, face_match_min_score: null }
      responses:
        '200':
          description: Semua setting setelah diubah
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettingsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422':
          description: 'Contoh: `patrol_location_radius_meters must be between 1 and 10000`, `max_speed: unknown setting`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
components:
  securitySchemes:
    AppId:
      type: apiKey
      in: header
      name: X-App-Id
      description: App ID aplikasi, contoh `sp_web_k3m9qx2ht5vbn4rw_w8xq2a`.
    AppKey:
      type: apiKey
      in: header
      name: X-App-Key
      description: |
        Khusus halaman dokumentasi. App Key dipakai browser untuk membuat `X-Signature`
        lalu dihapus dari request, sehingga tidak pernah terkirim ke server.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token dari `/auth/login`. Terisi otomatis setelah login di halaman ini.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    Search:
      name: search
      in: query
      schema: { type: string }
...
  responses:
...
    Success:
      description: Berhasil
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    BadRequest:
      description: Aksi tidak diizinkan untuk data ini
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: App signature atau token tidak valid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Role tidak memiliki akses
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Data tidak ditemukan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Data sudah dipakai (misalnya email atau kode role)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Input tidak valid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    Meta:
      type: object
      properties:
        message: { type: string, example: Login success }
        code: { type: integer, example: 200 }
        status: { type: string, example: success }
    Envelope:
      type: object
      properties:
        meta: { $ref: '#/components/schemas/Meta' }
        data: {}
    ErrorResponse:
      type: object
      properties:
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
          example: { message: Validation error, code: 422, status: Error }
        data:
          description: Detail error (string, daftar pesan validasi, atau null)
          oneOf:
            - type: string
            - type: array
              items: { type: string }
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 10 }
        total: { type: integer, example: 42 }
        total_pages: { type: integer, example: 5 }
    Role:
      type: object
      properties:
        id: { type: integer, format: int64, example: 5 }
        code: { type: string, example: security_team }
        name: { type: string, example: Tim Keamanan }
        description: { type: string, example: Petugas patroli keamanan di lapangan }
        is_system: { type: boolean, example: true }
        is_active: { type: boolean, example: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    RoleResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Role' }
    CreateRoleRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
          description: Huruf kecil, angka, atau underscore; diawali huruf; 3-50 karakter. Tidak bisa diubah.
          example: supervisor_shift
        name: { type: string, maxLength: 100, example: Supervisor Shift }
        description: { type: string, maxLength: 255 }
        is_active: { type: boolean, default: true }
    UpdateRoleRequest:
      type: object
      required: [name, is_active]
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 255 }
        is_active: { type: boolean }
...
    PatrolPoint:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        name: { type: string, example: Main Gate }
        location: { type: string, example: Building A - Front Entrance }
        nfc_code: { type: string, example: '04:A2:1F:9C' }
        latitude: { type: number, format: double, example: -6.2250138 }
        longitude: { type: number, format: double, example: 106.8008324 }
        is_location_match_required:
          type: boolean
          description: Petugas harus berada di koordinat titik ini saat check-in
        is_face_validation_required:
          type: boolean
          description: Petugas harus lolos validasi wajah saat check-in
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PatrolPointResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/PatrolPoint' }
    CreatePatrolPointRequest:
      type: object
      required: [name, location, nfc_code, latitude, longitude]
      properties:
        name: { type: string, maxLength: 150, example: Main Gate }
        location: { type: string, maxLength: 255, example: Building A - Front Entrance }
        nfc_code: { type: string, maxLength: 100, example: '04:A2:1F:9C' }
        latitude: { type: number, format: double, minimum: -90, maximum: 90, example: -6.2250138 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180, example: 106.8008324 }
        is_location_match_required: { type: boolean, default: false }
        is_face_validation_required: { type: boolean, default: false }
    UpdatePatrolPointRequest:
      type: object
      required: [name, location, nfc_code, latitude, longitude, is_location_match_required, is_face_validation_required]
      properties:
        name: { type: string, maxLength: 150 }
        location: { type: string, maxLength: 255 }
        nfc_code: { type: string, maxLength: 100 }
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
        is_location_match_required: { type: boolean }
        is_face_validation_required: { type: boolean }
...
    Setting:
      type: object
      properties:
        key: { type: string, example: patrol_location_radius_meters }
        group: { type: string, enum: [patrol, face, security] }
        type: { type: string, enum: [integer, number, boolean] }
        value: { description: Nilai saat ini (number atau boolean), example: 100 }
        default_value: { description: Nilai default, example: 100 }
        is_default: { type: boolean }
        min: { type: number, nullable: true, example: 1 }
        max: { type: number, nullable: true, example: 10000 }
        unit: { type: string, example: meters }
        description: { type: string }
        updated_by: { type: integer, format: int64, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
    SettingsResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Setting' }